Autobuild rank score default

+++ scripts/thesis/poorly_ranked_binder_sample/make_fake_pandda_for_high_ranking_non_hits.py
@@ -37,6 +37,8 @@ def try_make(path):
 
 
 def try_link(source_path, target_path):
+    if target_path.exists():
+        return
     try:
         os.symlink(source_path, target_path)
     except Exception as e:
